energy/requirements.txt
yfinance>=0.2,<1
pandas>=2.1,<3
numpy>=1.26,<2
requests>=2.31
python-dotenv>=1.0
websockets>=11
